+++ rust/dbn/src/enums.rs
@@ -1293,7 +1293,7 @@ pub enum TradingEvent {
     #[default]
     #[pyo3(name = "NONE")]
     None = 0,
-    /// Order entry and modification are not allowed.
+    /// Order entry is allowed. Modification and cancellation are not allowed.
     #[pyo3(name = "NO_CANCEL")]
     NoCancel = 1,
     /// A change of trading session occurred. Daily statistics are reset.